Refactor dataTable function into a component
