+++ widgets/demographics/statistics.js
@@ -1,6 +1,6 @@
 var R = require('react');
 var D = R.DOM;
-var U = require('../shared/domUtils.js');
+var DataTable = require('../dataTable');
 
 var Statistics = R.createClass({
   displayName: 'demographics-statistics',
@@ -10,7 +10,6 @@ var Statistics = R.createClass({
   },
 
   render: function() {
-    var data = this.props.popData;
     var fields = [
       ['allUsualResidents', 'Total number of residents'],
       ['personsPerHectare', 'Persons per hectare'],
@@ -20,10 +19,13 @@ var Statistics = R.createClass({
       ['communalEstablishmentResidents', 'Number of communal establishment residents']
     ];
 
+    var table = R.createElement(DataTable, { data: this.props.popData, fields: fields });
+
     var className = 'data-table col ' +
       (this.props.size === 'half' ? 'span_5_of_12' : 'span_12_of_12');
 
-    return D.div({ className: className }, U.dataTable(data, fields));
+
+    return D.div({ className: className }, table);
   }
 });
